requirements.txt
flask==2.3.3
requests==2.31.0
openai>=1.35.0
python-dotenv==1.0.0
redis==4.6.0
urllib3==2.0.4
validators==0.22.0 